Refactor: connectivity services relocation
